+++ nbmolviz/base/__init__.py
@@ -1,5 +1,4 @@
 from .base_widget import *
-from .drivers3d import *
 from .mdt2json import *
-from .widget2d import *
-from .widget3d import *
+from .molviz2d import *
+from .molviz3d import *
