+++ nbmolviz/methods/__init__.py
@@ -2,5 +2,4 @@ from .atoms import *
 from .atomgroups import *
 from .molecules import *
 from .method import *
-from .residues import *
 from .trajectory import *
